30.09

+++ src/01_Matrix/01_Matrix/TMatrix.h
@@ -8,7 +8,7 @@ template<typename ValType>
 class TMatrix : public TVector<TVector<ValType>>
 {
 public:
-    explicit TMatrix(size_t _size = 10);
+    explicit TMatrix(int _size = 10);
     TMatrix(const TMatrix&);
     TMatrix(const TVector<TVector<ValType> >&);
     ~TMatrix();
@@ -28,7 +28,7 @@ public:
 
     friend ostream& operator<<(ostream& out, const TMatrix& _v)
     {
-        for (size_t i = 0; i < _v.size; i++)
+        for (int i = 0; i < _v.size; i++)
         {
             out << _v.elems[i] << "\n";
         }
@@ -36,14 +36,14 @@ public:
     }
     friend istream& operator >> (istream& in, TMatrix& temp)
     {
-        for (size_t i = 0; i < temp.size; i++)
+        for (int i = 0; i < temp.size; i++)
             in >> temp.elems[i];
         return in;
     }
 };
 
 template<typename ValType>
-TMatrix<ValType>::TMatrix(size_t _size) : TVector<TVector<ValType> >(_size)
+TMatrix<ValType>::TMatrix(int _size) : TVector<TVector<ValType> >(_size)
 {
     for (int i = 0; i < _size; i++)
         elems[i] = TVector<ValType>(_size - i, i);
@@ -54,17 +54,17 @@ TMatrix<ValType>::TMatrix(const TMatrix<ValType>& temp)
 {
     size = temp.size;
     elems = new TVector<ValType>[temp.size];
-    for (size_t i = 0; i < temp.size; i++)
+    for (int i = 0; i < temp.size; i++)
         elems[i] = temp.elems[i];
 }
 
 template<typename ValType>
 TMatrix<ValType>::TMatrix(const TVector<TVector<ValType> >& temp) : TVector<TVector<ValType> >(temp.GetSize())
 {
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
     {
         elems[i] = TVector<ValType>(size - i, i);
-        for (size_t j = i; j < size; j++)
+        for (int j = i; j < size; j++)
             elems[i][j] = temp[i][j];
     }
 }
@@ -77,7 +77,7 @@ bool TMatrix<ValType>::operator==(const TMatrix<ValType>& temp) const
 {
     if (size != temp.size)
         throw "Размерности не совпадают";
-    for (size_t i = 0; i < temp.size; i++)
+    for (int i = 0; i < temp.size; i++)
     {
         if (elems[i] != temp.elems[i])
         {
@@ -113,7 +113,7 @@ template<typename ValType>
 TMatrix<ValType> TMatrix<ValType>::operator+(ValType temp)
 {
     TMatrix<ValType> rez(*this);
-    for (size_t i = 0; i < rez.size; i++)
+    for (int i = 0; i < rez.size; i++)
         rez.elems[i] = elems[i] + temp;
     return rez;
 }
@@ -122,7 +122,7 @@ template<typename ValType>
 TMatrix<ValType> TMatrix<ValType>::operator-(ValType temp)
 {
     TMatrix<ValType> rez(*this);
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
         rez.elems[i] = rez.elems[i] - temp;
     return rez;
 }
@@ -131,7 +131,7 @@ template<typename ValType>
 TMatrix<ValType> TMatrix<ValType>::operator*(ValType temp)
 {
     TMatrix<ValType> rez(*this);
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
         rez.elems[i] = rez.elems[i] * temp;
     return rez;
 }
@@ -142,7 +142,7 @@ TMatrix<ValType> TMatrix<ValType>::operator+(const TMatrix& temp)
     if (size != temp.size)
         throw "Размерности не совпадают";
     TMatrix<ValType> rez(*this);
-    for (size_t i = 0; i < rez.size; i++)
+    for (int i = 0; i < rez.size; i++)
         rez.elems[i] = elems[i] + temp.elems[i];
     return rez;
 }
@@ -153,7 +153,7 @@ TMatrix<ValType> TMatrix<ValType>::operator-(const TMatrix& temp)
     if (size != temp.size)
         throw "Размерности не совпадают";
     TMatrix<ValType> rez(*this);
-    for (size_t i = 0; i < rez.size; i++)
+    for (int i = 0; i < rez.size; i++)
         rez.elems[i] = elems[i] - temp.elems[i];
     return rez;
 }
@@ -164,14 +164,14 @@ TMatrix<ValType> TMatrix<ValType>::operator*(const TMatrix& temp)
     if (size != temp.size)
         throw "Размерности не совпадают";
     TMatrix<ValType> rez(*this);
-    for (size_t i = 0; i < rez.size; i++)
+    for (int i = 0; i < rez.size; i++)
     {
-        size_t cols = 1;
-        size_t si = temp.elems[i].GetStartIndex();
-        for (size_t j = si; j < rez.size; j++)
+        int cols = 1;
+        int si = temp.elems[i].GetStartIndex();
+        for (int j = si; j < rez.size; j++)
         {
             rez.elems[i][j] = ValType(0);
-            for (size_t k = 0; k < cols; k++)
+            for (int k = 0; k < cols; k++)
                 rez.elems[i][j] = rez.elems[i][j] + elems[i][k + i] * temp.elems[k + i][j];
             cols++;
         }
@@ -182,15 +182,15 @@ TMatrix<ValType> TMatrix<ValType>::operator*(const TMatrix& temp)
 template<typename ValType>
 TVector<ValType> TMatrix<ValType>::operator*(const TVector<ValType>& temp)
 {
-    size_t _size = temp.GetSize();
+    int _size = temp.GetSize();
     if (size != _size)
         throw "Размерности не совпадают";
     TVector<ValType> rez(_size);
-    for (size_t i = 0; i < _size; i++)
+    for (int i = 0; i < _size; i++)
     {
         rez[i] = ValType(0);
-        size_t si = elems[i].GetStartIndex();
-        for (size_t j = si; j < _size; j++)
+        int si = elems[i].GetStartIndex();
+        for (int j = si; j < _size; j++)
             rez[i] = rez[i] + elems[i][j] * temp[j];
     }
     return rez;
@@ -201,7 +201,7 @@ ValType TMatrix<ValType>::Determinant()
 {
     ValType res;
     res = ValType(1);
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
         res = res * elems[i][i];
     return res;
 }
